app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
paho-mqtt==2.1.0
aiosqlite==0.20.0
